substep devleopment
Current substep design won't do since substeps can have different timings that all play at their own pace. With this design all differently timed substeps would be forced to conform to one timing. The solution is to handle the substep or "autoactions" in the state subject itself. Lists or queues of auto actions can be set by the chapter nodes which run and are kept track of in the state subject.