sprite animation skipping implemented
Will use a generic struct to contain any type of sprite animation to cut down on the hashmaps needed for making each property configurable. Future additions of things like non-linear movement should be considered in the future (which can be implemented by messing with the time passed variable, although a new variable would need to be introduced to keep the original transition time constant).