refactoring and bug fix
list.erase(iterator) returns an iterator to next element. I had used that to set the iterator of the list again in the for loop, but then the iterator++ ran so the erase resulted in doubling up on the advance. Fixed by putting the iterator++ in the else condition when erase isn't modifying it. This was in the tickSpriteAnimations function in the state subject.