Sprite animation bug fix and view restructure
Sprite animations continued going if they passed the goal by a certain amount. This is fixed by setting it to the goal if it passes.
View file structure changed for easier navigation. The "game" folder needs further organization.
Added a BASE_PATH macro from cmake to allow asset paths to be for any system.