Start chapter node builder class
In the previous commit, the addition of the texture manager is for allowing the model be completely decoupled of objects that are related to view. This helps with revmoving dependance on opengl initialization.

Decided to go with builders to build and modify the different objects for the purposes of the engine. They would be used in the model subject and by the engine portion to modify the objects. The runtime/gameloop will not need them beyond for loading in the model.