Add every new node to node stray
Currently thinking about merging chapter and graph classes into one class.
Need to consider scene node loading from a serialization standpoint. Chapters can say what chapters follow them using IDs instead of pointers, to allow the model to load them only if/when needed.
Can a similar thing be done for nodes? This decision is mainly based on thinking more about how the model is serialized and loaded. This decision can be somewhat deferred (at least for the node level). But it will be assumed for now that chapters should not be loaded if not needed. Hence ID usage for tree construction is most likely to be more ideal.