engine node editor structure change
General interface defined in engine_node_action_fields but need to decide on how the action object themselves are stored in order to allow both new objects and editing existing ones with the same interface. Perhaps the actions can be dynamically allocated? Or the object can be static but that would be limiting.