ImGui/Freetype text crash bug fix
The issue was that the while loop that split the line of text into several lines of text would not exit when the text line was not finished and the size of the line was 0 due to a line being too long. Problem functions fittedScreenLines and computeBreakIndex.

The fix skips all remaining text if a line is 0 in length. Better fixes will be considered in the future such as putting in hyphens when words are too long (spanning more than a line?).