Engine character editor
the style followed here makes sure not to modify the model directly, and uses the getters on the object and setter with the builder. This is more troblesome than just accessing the data directly. But having things follow this interface allows for greater control, although it may be reassessed in the future.